Refactor code in abstract generator
